app/requirements.txt
channels==4.0.0
channels-redis==4.1.0
crispy-bootstrap5==2023.10
daphne==4.0.0
Django==5.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
django-encrypted-model-fields==0.6.5
django-filter==23.4
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
django-widget-tweaks==1.5.0
drf-yasg==1.21.7
geopy==2.4.1
gunicorn==21.2.0
paho-mqtt==1.6.1
python-dotenv==1.0.0
PyPDF2==3.0.1
psycopg2-binary==2.9.9
requests==2.31.0
whitenoise==6.6.0
